update bert multi-label
