+++ .gitignore
@@ -6,3 +6,5 @@ __pycache__/
 .idea
 .cache
 __pycache__
+*.pt
+*.pickle
