+++ classification/model/BERT/BERT.py
@@ -1,0 +1,42 @@
+import torch
+import numpy as np
+import gluonnlp as nlp
+
+from torch import nn
+from torch.utils.data import Dataset
+from kobert.utils import get_tokenizer
+from kobert.pytorch_kobert import get_pytorch_kobert_model
+
+
+class BERTDataset(Dataset):
+    def __init__(self, x, y, bert_tokenizer, max_len, pad, pair):
+        transform = nlp.data.BERTSentenceTransform(bert_tokenizer, max_seq_length=max_len, pad=pad, pair=pair)
+        self.sentences = [transform([i]) for i in x]
+        self.labels = [np.int32(i) for i in y]
+
+    def __getitem__(self, i):
+        return (self.sentences[i] + (self.labels[i], ))
+
+    def __len__(self):
+        return (len(self.labels))
+
+
+class BERTClassifier(nn.Module):
+    def __init__(self, bert, hidden_size=768, num_classes=None, dr_rate=None):
+        super(BERTClassifier, self).__init__()
+        self.bert = bert
+        self.dr_rate = dr_rate
+        self.classifier = nn.Linear(hidden_size , num_classes)
+        if dr_rate: self.dropout = nn.Dropout(p=dr_rate)
+    
+    def gen_attention_mask(self, token_ids, valid_length):
+        attention_mask = torch.zeros_like(token_ids)
+        for i, v in enumerate(valid_length):
+            attention_mask[i][:v] = 1
+        return attention_mask.float()
+
+    def forward(self, token_ids, valid_length, segment_ids):
+        attention_mask = self.gen_attention_mask(token_ids, valid_length)        
+        _, pooler = self.bert(input_ids = token_ids, token_type_ids = segment_ids.long(), attention_mask = attention_mask.float().to(token_ids.device), return_dict=False)
+        if self.dr_rate: out = self.dropout(pooler)
+        return self.classifier(out)
