+++ classification/model/BERT/kobert/utils/__init__.py
@@ -1,0 +1,1 @@
+from kobert.utils.utils import download, get_tokenizer
