--- classification/model/KoBERT/kobert/__init__.py
@@ -1,21 +1,0 @@
-# coding=utf-8
-# Copyright 2019 SK T-Brain Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from kobert.utils.utils import download, get_tokenizer
-from kobert.pytorch_kobert import get_pytorch_kobert_model
-from kobert.mxnet_kobert import get_mxnet_kobert_model
-from kobert.onnx_kobert import get_onnx_kobert_model
-
-__all__ = ("download", "get_tokenizer", "get_pytorch_kobert_model" ,"get_mxnet_kobert_model", "get_onnx_kobert_model")
