--- classification/model/KoBERT/kobert/utils/__init__.py
@@ -1,1 +1,0 @@
-from kobert.utils.utils import download, get_tokenizer
